Mc recommendation request form component (#82)
* mc - WIP two of 6 fields for commit form, plus story

Still tBD: finish other two fields on the form, write tests.

* mc - WIP fix formatting

* mc - WIP on commit form

need to add the Z for the timezone on create and edit.
also need to remove the Z that we get from the backend when filling in initial contents

* mc - strip off Z for timezone when populating commit form with exisiting commit

need to add it back in for create and update in another part of the codebase

* added tests for rr form, also fixed up formatting and changed date and done visu
als so that it defaults to an approvable value when updating

* fixed mutation coverage tests

---------

Co-authored-by: Ryan Hung <rh.lh.314@gmail.com>